(chore) Bunch of cleanup
