+++ containers/mosaic/src/components/common/index.ts
@@ -1,0 +1,1 @@
+export { FormDrawerWrapper } from "./FormDrawerWrapper";
