+++ containers/mosaic/src/utils/hooks/index.ts
@@ -2,3 +2,4 @@ export * from "./useScrollSpy";
 export { default as useScrollSpy } from "./useScrollSpy";
 
 export { default as useFieldErrors } from "./useFieldErrors";
+export { default as useMosaicSettings } from "./useMosaicSettings";
