+++ containers/mosaic/src/utils/index.ts
@@ -1,3 +1,1 @@
 export * from "./toggle";
-
-export { default as useMosaicSettings } from "./useMosaicSettings";
