+++ containers/sb-8/stories/components/ButtonRow/ButtonRow.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { ReactElement, useMemo } from "react";
 import styled from "styled-components";
-import { toggleMap, toggleOptions } from "@root/utils/storyUtils";
+import { toggleMap, toggleOptions } from "../../../utils";
 
 import ButtonRow from "@root/components/ButtonRow";
 import { ButtonProps } from "@root/components/Button";
