+++ containers/sb-8/stories/components/Field/FormFieldPhone/FormFieldPhone.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { ReactElement, useMemo } from "react";
 import { FieldDef } from "@root/components/Field";
 import Form, { useForm } from "@root/components/Form";
-import { renderButtons } from "@root/utils/storyUtils";
+import { renderButtons } from "../../../../utils";
 
 export default {
 	title: "FormFields/FormFieldPhone",
