+++ containers/sb-8/stories/components/Field/FormFieldToggle/FormFieldToggle.stories.tsx
@@ -4,7 +4,7 @@ import { ReactElement, useMemo } from "react";
 // Components
 import Form, { useForm } from "@root/components/Form";
 import { FieldDef } from "@root/components/Field";
-import { renderButtons } from "@root/utils/storyUtils";
+import { renderButtons } from "../../../../utils";
 
 export default {
 	title: "FormFields/FormFieldToggle",
