+++ containers/sb-8/stories/components/Form/Playground.stories.tsx
@@ -5,7 +5,7 @@ import { nanoid } from "nanoid";
 // Utils
 import { checkboxOptions } from "@root/components/Field/FormFieldCheckbox/FormFieldCheckboxUtils";
 import { useForm } from "@root/components/Form";
-import { renderButtons } from "@root/utils/storyUtils";
+import { renderButtons } from "../../../utils";
 
 // Components
 import Form from "@root/components/Form";
