+++ containers/sb-8/stories/components/Form/Profile.stories.tsx
@@ -6,7 +6,7 @@ import AddIcon from "@mui/icons-material/Add";
 // Utils
 import { useForm } from "@root/components/Form";
 import { validateEmail, validateSlow } from "@root/components/Form/validators";
-import { renderButtons, toggleMap, toggleOptions } from "@root/utils/storyUtils";
+import { renderButtons, toggleMap, toggleOptions } from "../../../utils";
 import { getOptionsCountries, getOptionsStates } from "@root/components/Field/FormFieldAddress/utils/optionGetters";
 
 // Components
