+++ containers/sb-8/stories/components/Form/Validators.stories.tsx
@@ -4,7 +4,7 @@ import { ReactElement, useEffect, useMemo } from "react";
 // Utils
 import { useForm } from "@root/components/Form";
 import { validateEmail, validateSlow, validateNumber, validateURL } from "@root/components/Form/validators";
-import { renderButtons } from "@root/utils/storyUtils";
+import { renderButtons } from "../../../utils";
 
 // Components
 import Form from "@root/components/Form";
