+++ containers/sb-8/stories/components/Form/VaryingSections.stories.tsx
@@ -3,7 +3,7 @@ import { ReactElement, useEffect, useMemo } from "react";
 
 // Utils
 import { useForm } from "@root/components/Form";
-import { renderButtons } from "@root/utils/storyUtils";
+import { renderButtons } from "../../../utils";
 
 // Components
 import Form from "@root/components/Form";
